feat: added tsdoc, removed unneeded dependencies
