+++ common/config/rush/repo-state.json
@@ -1,5 +1,5 @@
 // DO NOT MODIFY THIS FILE MANUALLY BUT DO COMMIT IT. It is generated and used by Rush.
 {
-  "pnpmShrinkwrapHash": "7e0cadb701aa8487beb407af331047529083bd4a",
+  "pnpmShrinkwrapHash": "ff8be416a7201750436ec92623a8d8361edf5745",
   "preferredVersionsHash": "bf21a9e8fbc5a3846fb05b4fa0859e0917b2202f"
 }
