+++ workbench-core/authorization/package.json
@@ -30,6 +30,7 @@
   },
   "dependencies": {
     "@aws/workbench-core-logging": "workspace:*",
+    "@aws/workbench-core-user-management": "workspace:*",
     "@casl/ability": "^5.4.3",
     "lodash": "^4.17.21"
   },
