--- workbench-core/authorization/src/dynamic/inputs/addUserToGroup.ts
@@ -1,8 +1,0 @@
-export interface AddUserToGroupRequest {
-  userId: string;
-  groupName: string;
-}
-
-export interface AddUserToGroupResponse {
-  added: boolean;
-}
