--- workbench-core/authorization/src/dynamic/inputs/createGroup.ts
@@ -1,7 +1,0 @@
-export interface CreateGroupRequest {
-  groupName: string;
-}
-
-export interface CreateGroupResponse {
-  created: boolean;
-}
