--- workbench-core/authorization/src/dynamic/inputs/deleteGroup.ts
@@ -1,7 +1,0 @@
-export interface DeleteGroupRequest {
-  groupName: string;
-}
-
-export interface DeleteGroupResponse {
-  deleted: boolean;
-}
