--- workbench-core/authorization/src/dynamic/inputs/doesGroupExist.ts
@@ -1,7 +1,0 @@
-export interface DoesGroupExistRequest {
-  groupName: string;
-}
-
-export interface DoesGroupExistResponse {
-  exists: boolean;
-}
