--- workbench-core/authorization/src/dynamic/inputs/getGroupUsers.ts
@@ -1,7 +1,0 @@
-export interface GetGroupUsersRequest {
-  groupName: string;
-}
-
-export interface GetGroupUsersResponse {
-  userIds: string[];
-}
