--- workbench-core/authorization/src/dynamic/inputs/getUserGroups.ts
@@ -1,7 +1,0 @@
-export interface GetUserGroupsRequest {
-  groupName: string;
-}
-
-export interface GetUserGroupsResponse {
-  groupIds: string[];
-}
