--- workbench-core/authorization/src/dynamic/inputs/isUserAssignedToGroup.ts
@@ -1,8 +1,0 @@
-export interface IsUserAssignedToGroupRequest {
-  userId: string;
-  groupName: string;
-}
-
-export interface IsUserAssignedToGroupResponse {
-  assigned: boolean;
-}
