--- workbench-core/authorization/src/dynamic/inputs/removeUserFromGroup.ts
@@ -1,8 +1,0 @@
-export interface RemoveUserFromGroupRequest {
-  userId: string;
-  groupName: string;
-}
-
-export interface RemoveUserFromGroupResponse {
-  removed: boolean;
-}
