--- workbench-core/authorization/src/dynamic/inputs/setGroupStatus.ts
@@ -1,8 +1,0 @@
-export interface SetGroupStatusRequest {
-  groupName: string;
-  status: string;
-}
-
-export interface SetGroupStatusResponse {
-  statusSet: boolean;
-}
