+++ workbench-core/authorization/src/dynamicAuthorization/dynamicAuthorizationInputs/setGroupStatus.ts
@@ -1,0 +1,22 @@
+/**
+ * Request object for SetGroupStatus
+ */
+export interface SetGroupStatusRequest {
+  /**
+   * Group id associated to the group the status is being set on
+   */
+  groupId: string;
+  /**
+   * Status to set
+   */
+  status: string;
+}
+/**
+ * Response object for SetGroupStatus
+ */
+export interface SetGroupStatusResponse {
+  /**
+   * States whether the status was successfully set
+   */
+  statusSet: boolean;
+}
