+++ workbench-core/base/package.json
@@ -58,8 +58,6 @@
     "@aws-sdk/s3-request-presigner": "^3.212.0",
     "@aws-sdk/types": "^3.212.0",
     "@aws-sdk/util-dynamodb": "^3.212.0",
-    "@aws/workbench-core-audit": "workspace:*",
-    "@aws/workbench-core-logging": "workspace:*",
     "@hapi/boom": "^10.0.0",
     "aws-sdk-client-mock": "^0.6.2",
     "jsonschema": "^1.4.1",
